App Layout component added to its own file
